+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - [stylelint] Empower runner [#420](https://github.com/sider/runners/pull/420)
 - [cpplint] New support [#417](https://github.com/sider/runners/pull/417)
 - Add Options module and delete insecure options [#415](https://github.com/sider/runners/pull/415)
+- Output an unexpected error message to the trace log [#446](https://github.com/sider/runners/pull/446)
 
 ## 0.7.5